moved funktion logic to separate file
