+++ web_scrapers/scrape.sh
@@ -4,4 +4,5 @@ python3 scrape_zutrittsberechtigte.py
 echo ""
 echo "writing delta.sql..."
 python3 create_zutrittsberechtigte_update.py > delta.sql
+deactivate
 echo "done!"
